nodeStatus renamed to get-node-status
